create a new hash2group crate
