+++ .github/workflows/docs.yml
@@ -33,7 +33,7 @@ jobs:
       - name: install toolchain
         uses: dtolnay/rust-toolchain@nightly
       - name: cargo doc
-        run: cargo doc --all-features
+        run: cargo doc --workspace --all-features
       - name: page configuration
         uses: actions/configure-pages@v3
       - name: Upload artifact
