+++ .github/workflows/pr.yml
@@ -50,9 +50,9 @@ jobs:
       - name: Run clippy
         run: |
           case "${{ matrix.features }}" in
-            default)    cargo clippy --all-targets -- -D warnings ;;
-            no-default) cargo clippy --all-targets --no-default-features -- -D warnings ;;
-            all)        cargo clippy --all-targets --all-features -- -D warnings ;;
+            default)    cargo clippy --workspace --all-targets -- -D warnings ;;
+            no-default) cargo clippy --workspace --all-targets --no-default-features -- -D warnings ;;
+            all)        cargo clippy --workspace --all-targets --all-features -- -D warnings ;;
           esac
 
   test:
@@ -71,16 +71,16 @@ jobs:
       - name: Build
         run: |
           case "${{ matrix.features }}" in
-            default)    cargo build --verbose ;;
-            no-default) cargo build --no-default-features --verbose ;;
-            all)        cargo build --all-features --verbose ;;
+            default)    cargo build --workspace --verbose ;;
+            no-default) cargo build --workspace --no-default-features --verbose ;;
+            all)        cargo build --workspace --all-features --verbose ;;
           esac
       - name: Test
         run: |
           case "${{ matrix.features }}" in
-            default)    cargo test --verbose --no-fail-fast ;;
-            no-default) cargo test --no-default-features --verbose --no-fail-fast ;;
-            all)        cargo test --all-features --verbose --no-fail-fast ;;
+            default)    cargo test --workspace --verbose --no-fail-fast ;;
+            no-default) cargo test --workspace --no-default-features --verbose --no-fail-fast ;;
+            all)        cargo test --workspace --all-features --verbose --no-fail-fast ;;
           esac
 
   wasm:
@@ -90,9 +90,9 @@ jobs:
       - uses: actions/checkout@v6
       - uses: bytecodealliance/actions/wasmtime/setup@v1
       - name: Build
-        run: cargo build --target wasm32-wasip1 --verbose
+        run: cargo build --workspace --target wasm32-wasip1 --verbose
       - name: Test
-        run: cargo test --target wasm32-wasip1 --verbose --no-fail-fast
+        run: cargo test --workspace --target wasm32-wasip1 --verbose --no-fail-fast
 
   bench:
     name: Benchmark smoke test
@@ -111,7 +111,7 @@ jobs:
     steps:
       - uses: actions/checkout@v6
       - uses: dtolnay/rust-toolchain@nightly
-      - run: cargo doc --all-features --no-deps
+      - run: cargo doc --workspace --all-features --no-deps
 
   # Anchor job for branch protection. Point required status checks at this.
   # see: https://github.com/orgs/community/discussions/26822
