+++ Cargo.toml
@@ -1,3 +1,4 @@
+
 [package]
 name = "sigma-proofs"
 version = "0.3.0"
@@ -8,15 +9,34 @@ authors = [
     "Victor Snyder-Graf <victor@subgraf.dev>",
     "Lénaïck Gouriou <lg@leanear.io>",
 ]
-edition = "2021"
-license = "MIT OR Apache-2.0"
+edition.workspace = true
+license.workspace = true
 readme = "README.md"
-repository = "https://github.com/sigma-rs/sigma-proofs"
+repository.workspace = true
 documentation = "https://docs.rs/sigma-proofs"
-categories = ["cryptography"]
+categories.workspace = true
 keywords = ["cryptography", "zero-knowledge", "NIZK", "sigma-protocols"]
 description = "A toolkit for auto-generated implementations of Σ-protocols"
 exclude = [".gitignore"]
+[workspace]
+members = [".", "hash2group"]
+
+[workspace.package]
+edition = "2021"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/sigma-rs/sigma-proofs"
+categories = ["cryptography"]
+
+[workspace.dependencies]
+bytemuck = { version = "1.25.0", default-features = false }
+curve25519-dalek = { version = "4.1.3", default-features = false }
+digest = { version = "0.11.2", default-features = false }
+elliptic-curve = { version = "0.13.8", default-features = false }
+hex-literal = "0.4"
+k256 = { version = "0.13.4", default-features = false }
+p256 = { version = "0.13", default-features = false }
+sha2 = { version = "0.11", default-features = false }
+sha3 = { version = "0.11.0", default-features = false }
 
 [features]
 # Include commonly used curved in default features.
@@ -38,22 +58,19 @@ p256 = ["dep:p256", "spongefish/p256"]
 [dependencies]
 ahash = { version = "0.8", default-features = false }
 bls12_381 = { version = "0.8.0", default-features = false, features = ["groups"], optional = true }
-# NOTE: min_const_generics feature ensures that the Zeroable trait is implemented for all [u8; _].
-bytemuck = { version = "1.25.0", features = ["min_const_generics"] }
-curve25519-dalek = { version = "4.1.3", default-features = false, features = ["alloc", "group"], optional = true }
-digest = "0.11.2"
-elliptic-curve = { version = "0.13.8", features = ["hash2curve"] }
+bytemuck = { workspace = true }
+curve25519-dalek = { workspace = true, features = ["alloc", "group"], optional = true }
 ff = { version = "0.13", features = ["derive"] }
 group = "0.13.0"
 hashbrown = { version = "0.15", default-features = false }
-k256 = { version = "0.13.4", default-features = false, features = ["arithmetic", "alloc", "hash2curve", "expose-field"], optional = true }
+k256 = { workspace = true, features = ["arithmetic", "alloc", "expose-field"], optional = true }
 keccak = { version = "0.1.5", default-features = false }
 num-bigint = { version = "0.4.6", default-features = false }
 num-traits = { version = "0.2.19", default-features = false, features = ["libm"] }
-p256 = { version = "0.13", default-features = false, features = ["arithmetic", "expose-field", "hash2curve"], optional = true }
+p256 = { workspace = true, features = ["arithmetic", "expose-field"], optional = true }
 rand = { version = "0.8.5", optional = true }
 rand_core = { version = "0.6", default-features = false }
-sha3 = { version = "0.11.0", default-features = false }
+sha3 = { workspace = true }
 spongefish = { version = "0.5", features = ["derive"] }
 subtle = { version = "2.6.1", default-features = false }
 thiserror = { version = "1", optional = true }
@@ -63,20 +80,20 @@ zeroize = { version = "1.8.1", default-features = false, features = ["alloc"] }
 [dev-dependencies]
 anyhow = "1.0.100"
 bls12_381 = "0.8.0"
-curve25519-dalek = { version = "4", default-features = false, features = ["serde", "rand_core", "alloc", "digest", "precomputed-tables", "group"] }
+curve25519-dalek = { workspace = true, features = ["serde", "rand_core", "alloc", "digest", "precomputed-tables", "group"] }
 divan = "0.1.21"
 hex = "0.4"
-hex-literal = "0.4"
+hex-literal = { workspace = true }
 json = "0.12.4"
-k256 = { version = "0.13", features = ["arithmetic"] }
+k256 = { workspace = true, features = ["arithmetic"] }
 libtest-mimic = "0.8.1"
-p256 = { version = "0.13", features = ["arithmetic"] }
+p256 = { workspace = true, features = ["arithmetic"] }
 rand_chacha = { version = "0.3.1" }
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
 serde_with = { version = "3.16.1", features = ["hex"] }
 serial_test = "3.3.1"
-sha2 = "0.11"
+sha2 = { workspace = true }
 spongefish = { version = "0.5", features = ["curve25519-dalek", "p256", "bls12_381", "derive", "keccak", "std"] }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
